requirements.txt
Markdown==3.10.2
Pillow==12.1.1
